// src/app/api/moldb/v1/model/dataDfKcat.ts
/**
 * PostgREST API
 * This is a dynamic API generated by PostgREST
 *
 * The version of the OpenAPI document: 12.2.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DataDfKcat { 
    ec?: string;
    substrate?: string;
    organism?: string;
    uniprot?: string;
    enzymetype?: string;
    ph?: number;
    temperature?: number;
    pubmedid?: number;
    KCAT_VALUE?: number;
    unit?: string;
    smiles?: string;
}